implemented basic user policy

+++ app/controllers/users_controller.rb
@@ -1,25 +1,30 @@
 class UsersController < ApplicationController
-  # before_filter :authenticate_user!
-  # after_action :verify_authorized
+  before_filter :authenticate_user!
+  after_action :verify_authorized
 
   def index
     @users = User.all
-    # authorize User
+
+    authorize User.new
   end
 
   def show
     @user = User.find(params[:id])
-    # authorize @user
+
+    authorize @user
   end
 
   def edit
     @user = User.find(params[:id])
-    # authorize @user
+
+    authorize @user
   end
 
   def update
     @user = User.find(params[:id])
-    # authorize @user
+
+    authorize @user
+
     if @user.update_attributes(secure_params)
       redirect_to admin_index_path, :notice => "User #{@user.name} was succesfully updated."
     else
@@ -29,7 +34,7 @@ class UsersController < ApplicationController
 
   def destroy
     user = User.find(params[:id])
-    # authorize user
+    authorize user
     user.destroy
     redirect_to users_path, :notice => "User deleted."
   end
